Former-commit-id: 12f4d245c118bce8135b443875c6fd2d91c18045 [formerly 4979f94478058a94ff3881f74e24a55949ee6931] [formerly 8d16d2536ebde1d6e5ae23f73f2bb3875d0cf748 [formerly 0deb1b21d4acdbac6a9276cc1794645ee0bb8ccb]]
Former-commit-id: 8d16d2536ebde1d6e5ae23f73f2bb3875d0cf748
Former-commit-id: c0a1b904ad1dda08e9ea0bd742c9db01ce785cab

+++ backend/backtest/models.py
@@ -27,7 +27,7 @@ class Result(models.Model):
     mdd = models.FloatField(blank=True, null=True)
     win_lose_rate = models.FloatField(blank=True, null=True)
     final_asset = models.BigIntegerField(blank=True, null=True)
-    final_earn = models.IntegerField(blank=True, null=True)
+    final_earn = models.BigIntegerField(blank=True, null=True)
     final_rate = models.FloatField(blank=True, null=True)
     created_at = models.DateTimeField(auto_now_add=True)
     
@@ -64,7 +64,7 @@ class DayHistory(models.Model):
     result = models.ForeignKey(Result, on_delete=models.CASCADE)
     date = models.CharField(max_length=20)
     day_earn_rate = models.FloatField()
-    day_earn = models.IntegerField()
+    day_earn = models.BigIntegerField()
     current_asset = models.BigIntegerField()
     
     class Meta:
